requirements.txt
absl-py==1.4.0
addict==2.4.0
anyio==3.7.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-lru==2.0.2
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.7.0
boltons @ file:///croot/boltons_1677628692245/work
cachetools==5.3.1
causal-conv1d @ file:///hy-tmp/causal_conv1d-1.0.0%2Bcu118torch1.13cxx11abiFALSE-cp38-cp38-linux_x86_64.whl
certifi @ file:///croot/certifi_1683875369620/work/certifi
cffi @ file:///tmp/abs_98z5h56wf8/croots/recipe/cffi_1659598650955/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
chumpy==0.70
click==8.1.7
comm==0.1.3
conda @ file:///croot/conda_1685025188142/work
conda-content-trust @ file:///tmp/abs_5952f1c8-355c-4855-ad2e-538535021ba5h26t22e5/croots/recipe/conda-content-trust_1658126371814/work
conda-package-handling @ file:///croot/conda-package-handling_1666940373510/work
ConfigArgParse==1.7
contourpy==1.0.7
cryptography @ file:///croot/cryptography_1677533068310/work
cycler==0.11.0
Cython==3.0.9
daal==2023.1.1
daal4py==2023.1.1
dash==2.16.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
einops==0.7.0
exceptiongroup==1.1.1
executing==1.2.0
fastjsonschema==2.17.1
filelock==3.13.3
Flask==3.0.2
fonttools==4.39.4
fqdn==1.5.1
freetype-py==2.4.0
fsspec==2024.3.1
google-auth==2.19.0
google-auth-oauthlib==1.0.0
grad-cam==1.5.0
grpcio==1.54.2
huggingface-hub==0.22.2
idna @ file:///croot/idna_1666125576474/work
imageio==2.34.0
importlib-metadata==6.6.0
importlib-resources==5.12.0
ipykernel==6.23.1
ipython==8.12.2
ipython-genutils==0.2.0
ipywidgets==8.1.2
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.14
jsonpatch @ file:///tmp/build/80754af9/jsonpatch_1615747632069/work
jsonpointer==2.1
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter-lsp==2.2.0
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyter_server==2.6.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.0
jupyterlab-language-pack-zh-CN==4.0.post0
jupyterlab-pygments==0.2.2
jupyterlab_server==2.22.1
jupyterlab_widgets==3.0.10
keras==2.12.0
kiwisolver==1.4.4
lazy_loader==0.3
llvmlite==0.41.1
mamba-ssm @ file:///hy-tmp/mamba_ssm-1.0.1%2Bcu118torch1.13cxx11abiFALSE-cp38-cp38-linux_x86_64.whl
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mistune==2.0.5
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.4.0
nbformat==5.8.0
nest-asyncio==1.5.6
networkx==3.1
ninja==1.11.1.1
notebook==6.5.4
notebook_shim==0.2.3
numba==0.58.1
numpy==1.24.3
nvidia-ml-py==11.525.112
nvitop==1.1.2
oauthlib==3.2.2
open3d==0.18.0
opencv-python==4.9.0.80
opendr==0.78
overrides==7.3.1
packaging @ file:///croot/packaging_1678965309396/work
pandas==2.0.2
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
platformdirs==3.5.1
plotly==5.20.0
pluggy @ file:///tmp/build/80754af9/pluggy_1648042571233/work
plyfile==1.0.3
prometheus-client==0.17.0
prompt-toolkit==3.0.38
protobuf==4.23.2
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycocotools==2.0.7
pycosat @ file:///croot/pycosat_1666805502580/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pyglet==2.0.14
Pygments==2.15.1
PyOpenGL==3.1.0
pyOpenSSL @ file:///croot/pyopenssl_1677607685877/work
pyparsing==3.0.9
pyquaternion==0.9.9
pyrender==0.1.45
pyrsistent==0.19.3
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.1.0
regex==2023.12.25
requests @ file:///croot/requests_1682607517574/work
requests-oauthlib==1.3.1
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.9
ruamel.yaml @ file:///croot/ruamel.yaml_1666304550667/work
ruamel.yaml.clib @ file:///croot/ruamel.yaml.clib_1666302247304/work
safetensors==0.4.2
scikit-image==0.21.0
scikit-learn==1.2.2
scikit-learn-intelex==2023.1.1
scipy==1.10.1
Send2Trash==1.8.2
six @ file:///tmp/build/80754af9/six_1644875935023/work
sklearn==0.0.post5
sniffio==1.3.0
soupsieve==2.4.1
stack-data==0.6.2
tbb==2021.9.0
tenacity==8.2.3
tensorboard==2.13.0
tensorboard-data-server==0.7.0
termcolor==2.3.0
terminado==0.17.1
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tifffile==2023.7.10
timm==0.4.12
tinycss2==1.2.1
tokenizers==0.15.2
tomli==2.0.1
toolz @ file:///croot/toolz_1667464077321/work
torch==1.13.0+cu117
torchaudio==0.13.0+cu117
torchvision==0.14.0+cu117
tornado==6.3.2
tqdm @ file:///croot/tqdm_1679561862951/work
traitlets==5.9.0
transformers==4.39.2
transforms3d==0.4.1
trimesh==4.2.4
triton==2.2.0
ttach==0.0.3
typing_extensions==4.6.2
tzdata==2023.3
uri-template==1.2.0
urllib3==1.25.8
visualize==0.5.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.2
Werkzeug==3.0.1
widgetsnbextension==4.0.10
xgboost==1.7.5
zipp==3.15.0
